// src/components/types/CarRecord.ts
export interface CarRecord {
  id: string;
  ModelCode: string;
  Make: string;
  Model: string;
  Description: string | null;
  CleanCarFeebate: string;
  RRPincGST: string;
  _48_50000km: string;
  _48_40000km: string;
  _48_30000km: string;
  _48_20000km: string;
  _48_10000km: string;
  _36_50000km: string;
  _36_40000km: string;
  _36_20000km: string;
  _36_30000km: string;
  _36_10000km: string;
  _24_50000km: string;
  _24_40000km: string;
  _24_30000km: string;
  _24_20000km: string;
  _24_10000km: string;
  _12_40000km: string;
  _12_50000km: string;
  _12_30000km: string;
  _12_10000km: string;
  _12_20000km: string;
}